% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_shortcuts}
\alias{list_shortcuts}
\title{List Shortcuts and Their Current Key Bindings}
\usage{
list_shortcuts(path_shortcuts = NULL, path_addins_json = NULL)
}
\arguments{
\item{path_shortcuts}{Path to \code{.shrtcts.R} or \code{.shrtcts.yml}. See \code{\link[=locate_shortcuts_source]{locate_shortcuts_source()}} for more detail.}

\item{path_addins_json}{Path to RStudio's \code{addins.json} keybindings config file. See \code{\link[=locate_addins_json]{locate_addins_json()}} for more detail.}
}
\value{
A data frame with the shortcut \code{name}, it's assigned \code{addin}
placeholder function, the \code{shrcts_keybinding} declared in the shrtcts
source file, and the \code{rstudio_keybinding} currently assigned to the
shortcut in RStudio.
}
\description{
Lists all shortcuts declared in \code{.shrtcts.R} or \code{.shrtcts.yml}. Also lists
the currently assigned RStudio keyboard shortcuts.
}
\examples{
if (
  interactive() &&
  requireNamespace("rstudioapi", quietly = TRUE) &&
  rstudioapi::hasFun("versionInfo")
 ) {
  list_shortcuts()
}

}
\seealso{
\code{\link[=add_rstudio_shortcuts]{add_rstudio_shortcuts()}}, \link{paths}
}
