% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{paths}
\alias{paths}
\alias{locate_shortcuts_source}
\alias{locate_addins_json}
\title{Paths Used by shrtcts}
\usage{
locate_shortcuts_source(path = NULL, all = FALSE)

locate_addins_json(path = NULL, all = FALSE)
}
\arguments{
\item{path}{The path to the file.}

\item{all}{List all found files, otherwise only the first is returned.}
}
\description{
These functions help locate the paths used by \pkg{shrtcts} when looking for
configuration files. You can use these functions to verify that \pkg{shrtcts}
is locating the files where you expect them to be. See the \strong{Options}
section below for global options that can be used to override the default
behavior of these functions.
}
\section{Functions}{
\itemize{
\item \code{locate_shortcuts_source()}: Find the path to \code{.shrtcts.R} or \code{.shrtcts.yml}.

\item \code{locate_addins_json()}: Find the path to \code{addins.json}, the RStudio configuration
file containing keyboard shortcuts for addins

}}
\section{Options}{


Use the following global options to set the default values for the
following files:
\itemize{
\item \code{shrtcts.path}: The path to \code{.shrtrcts.R} or \code{.shrtcts.yml}
\item \code{shrtcts.addins_json}: The path to the RStudio \code{addins.json} file used for
setting keyboard shortcuts for addins.
}
}

\examples{
if (interactive()) {
  locate_shortcuts_source()
  locate_addins_json()
}

}
