% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{shortcut_run}
\alias{shortcut_run}
\alias{shortcut_get}
\title{Get and run shortcuts by name}
\usage{
shortcut_run(.name, ..., .path_shortcuts = NULL)

shortcut_get(.name, .path_shortcuts = NULL)
}
\arguments{
\item{.name}{Name of the function of the shortcut (not from the roxygen2
comments but from the name to which the function is assigned).}

\item{...}{Additional arguments passed to the shortcut function when run.}

\item{.path_shortcuts}{The path to your \code{.shrcts.R} file. Must be an R
script; the YAML shortcuts syntax is not supported.}
}
\description{
If you're using an \code{.R} file to store your shortcuts, and if your shortcuts
use named functions, you can get and run the shortcut functions with these
helpers.

For example, if your \verb{~/.shrtcts.R} file includes a \code{cow_say_praise()}
function

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# ~/.shrtcts.R

praise <- function() \{
  # your shortcuts, when run this way, can reference other functions
  praise::praise()
\}

#' Have a cow say something nice
cow_say_praise <- function() \{
  cowsay::say(praise())
\}
}\if{html}{\out{</div>}}

you can use

\if{html}{\out{<div class="sourceCode r">}}\preformatted{shortcut_run("cow_say_praise")
}\if{html}{\out{</div>}}

to run the shortcut, or you can
use

\if{html}{\out{<div class="sourceCode r">}}\preformatted{happy_cow <- shortcut_get("cow_say_praise")
}\if{html}{\out{</div>}}

to get the shortcut function to run yourself.

With these helper functions, your \verb{~/.shrtcts.R} file is more portable and
becomes a place you can store utility or workflow functions.
\subsection{Using shortcuts in Positron}{

\code{shortcut_run()} enables you to use your shortcuts in Positron. You can
assign a keyboard shortcut to a specific shortcut function by running the
\emph{Preferences: Open Default Keyboard Shortcuts (JSON)} command. Then add an
entry like the following:

\if{html}{\out{<div class="sourceCode json">}}\preformatted{\{
  "key": "cmd+k cmd+p",
  "name": "Have a cow say something nice",
  "command": "workbench.action.executeCode.console",
  "args": \{
    "langId": "r",
    "code": "shrtcts::shortcut_run('cow_say_praise')",
    "focus": true
  \}
\},
}\if{html}{\out{</div>}}

When you press \code{Cmd + K} followed by \code{Cmd + P}, Positron will run the
\code{cow_say_praise} shortcut in your R console. Replace
\code{workbench.action.executeCode.console} with
\code{workbench.action.executeCode.silently} to run the shortcut silently in the
background. You can also use the \code{"when"} key to control when the keyboard
shortcut is valid, e.g. \code{"when": "editorTextFocus"}.
}
}
\section{Functions}{
\itemize{
\item \code{shortcut_run()}: Run a shortcut by assigned name

\item \code{shortcut_get()}: Get a shortcut by assigned name

}}
\examples{
# Create a small example .shrtcts.R file.
set.seed(42)
tmp_shortcuts <- tempfile(fileext = ".R")
writeLines("
praise <- function() {
  # your shortcuts, when run this way, can reference other functions
  praise::praise()
}

#' Have a cow say something nice
cow_say_praise <- function() {
  cowsay::say(praise())
}
", tmp_shortcuts)

# Run the `cow_say_praise` shortcut by name
shortcut_run("cow_say_praise", .path_shortcuts = tmp_shortcuts)

# Get the `cow_say_praise` shortcut
cow_praise <- shortcut_get("cow_say_praise", tmp_shortcuts)
cow_praise()

}
