% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrtcts-package.R
\docType{package}
\name{shrtcts-package}
\alias{shrtcts}
\alias{shrtcts-package}
\title{shrtcts: Make Anything an RStudio Shortcut}
\description{
An easily customizable method for creating RStudio
    addins from arbitrary R code. Use your own functions or call functions
    from other packages. Your shortcuts are stored in a YAML file in your
    home directory.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkg.garrickadenbuie.com/shrtcts}
  \item \url{https://github.com/gadenbuie/shrtcts}
  \item Report bugs at \url{https://github.com/gadenbuie/shrtcts/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{garrick@adenbuie.com} (\href{https://orcid.org/0000-0002-7111-0077}{ORCID})

}
\keyword{internal}
